Minor code tidying
Moved code that doesn't actually include a header (is entirely
functional) to the php dir so it's not a root file, tried to seperate
login stuff out to a php-auth folder etc but the include path to the
header changed depending on what directory you were in so i sacked it.
Also moved the mysql access stuff to the header to stop repeated
defenitions of it.